Thursday Lab (13/7/2023)
Lab -1
Write a Java program that prompts the user to enter a positive integer and calculates its factorial using a while loop. The program should display the factorial as the output.
Lab-2
Write a Java program that prompts the user to enter a number and then prints the multiplication table of that number using a for loop. The table should display multiplication from 1 to 10.