Thursday Lab (6/7/2023)
1.Create a class called "Person" with private instance variables "name" and "age".
 Implement getter and setter methods for both variables. 
Write a main method to create an instance of the class 
and demonstrate the encapsulation by accessing and modifying the variables using the getter and setter methods.
2.Create a class called "Person" with instance variables "name" and "age". 
Implement getter and setter methods for both variables. 
Create a subclass called "Employee" that inherits from "Person" 
and adds an instance variable called "salary". 
Implement getter and setter methods for the "salary" variable. 
Finally, create another subclass called "Manager" that inherits from "Employee" 
and adds an instance variable called "department". 
Implement getter and setter methods for the "department" variable. 
Write a main method to create an instance of the "Manager" class, 
demonstrate the inheritance by accessing and modifying the variables using the getter and setter methods at each level of the hierarchy.
3.Create a class called "Animal" with instance variables "name" and "age". 
Implement getter and setter methods for both variables. 
Create a subclass called "Dog" that inherits from "Animal" 
and adds an instance variable called "breed". 
Implement getter and setter methods for the "breed" variable. 
Write a main method to create an instance of the "Dog" class, 
demonstrate the inheritance by accessing and modifying the variables using the getter and setter methods.