Continue Statement 18/7/2023
Exercise 1:
Write a program that prompts the user to enter a series of integers. The program should calculate the sum of all the even numbers entered, skipping any negative numbers. Use the continue statement to skip negative numbers. 
Exercise : 2
Write a program that prints the Fibonacci series up to a specified limit entered by the user. However, skip printing numbers that are divisible by 3. Use the continue statement to skip such numbers.