Arithmetic, Logical, Assignment Operators 19/7/23
1. Arithmetic Operators:
Program where user is prompted to enter two numbers, which are stored in the variables num1 and num2.Then, we perform various arithmetic operations on these numbers, such as addition, subtraction, multiplication, division, and modulus.
2.Logical Operators:
Program where the user is prompted to enter a number, which is stored in the variable num. Then, we apply logical operators to check various conditions on this number.
3. Assignment Operators
The user is prompted to enter the initial value,  increment value, and decrement value. These values are stored in the variables value, increment, and decrement, respectively.