Comparison Operators  20/6/2023
Create a program to use the Scanner class to read user input from the console. The user is prompted to enter two numbers, which are stored in the variables num1 and num2.
We then use various comparison operators (==, !=, >, <, >=, <=) to compare these two numbers and store the results in boolean variables (isEqual, isNotEqual, isGreater, isLess, isGreaterOrEqual, isLessOrEqual).