Thursday Lab 3/8/2023
Question 1: Write a Java program to find the average of all elements in an array of doubles.
Question 2: Given an array of integers, write a Java program to find the maximum and minimum elements in the array.
Question 3: Create a superclass called "Shape" with properties like "numSides" and "color". Create three subclasses "Triangle", "Square", and "Circle" that inherit from "Shape" and have additional properties like "sideLengths" for triangle and square, and "radius" for the circle. Write a Java program to display the details of a triangle, a square, and a circle.
Question 4: Create an abstract class called "Vehicle" with abstract methods "start()" and "stop()". Create two subclasses "Car" and "Motorcycle" that extend the "Vehicle" class and implement the abstract methods. Write a Java program to demonstrate starting and stopping operations for a car and a motorcycle.